Create parse_bytes.py
Parsing Bytes 
Function called parse_bytes  that accepts a single string.
It should return a list of the binary bytes contained in the string.  Each byte is just a combination of eight 1's or 0's. For example:

parse_bytes("11010101 101 323")    # ['11010101']

parse_bytes("my data is: 10101010 11100010")    # ['10101010', '11100010']

parse_bytes("asdsa")   # []

Not all bytes will have a space before and after, some come at the beginning of a file or at the end.