Create next_prime.py
Function called next_prime, which returns a generator that will yield an unlimited number of primes, starting from the first prime (2).
Recall that a prime number is any whole number that has exactly two divisors: one and the number itself. The first few primes are 2, 3, 5, 7, 11, ... .


Example: Using a for...else statement to find prime number

num = 407

# To take input from the user
#num = int(input("Enter a number: "))

if num == 1:
    print(num, "is not a prime number")
elif num > 1:
   # check for factors
   for i in range(2,num):
       if (num % i) == 0:
           print(num,"is not a prime number")
           print(i,"times",num//i,"is",num)
           break
   else:
       print(num,"is a prime number")
       
# if input number is less than
# or equal to 1, it is not prime
else:
   print(num,"is not a prime number")


Output

407 is not a prime number
11 times 37 is 407

Here, we have used a for..else statement to check if num is prime.

It works on the logic that the else clause of the for loop runs if and only if we don't break out the for loop. That condition is met only when no factors are found, which means that the given number is prime.

So, in the else clause, we print that the number is prime.