Create main.py
This code perform scrapping of zillow.com site by using BeautifulSoup and Selenium to send results of search from zillow in google form for further analysis. 
It based that zillow stores data of the apartments in scripts tags in json format.

In google form it saves a link to an apartment, its price and address.

In case zillow changes its site to prevent scrapping to to print json data to file for clearer overview of it to fix dict keys for data extract.